refactor: move challenge creation into catalog bounded context
