--- backend/challenge-service/src/main/java/com/itachallenges/challengeservice/application/dto/CreateChallengeCommand.java
@@ -1,3 +1,0 @@
-package com.itachallenges.challengeservice.application.dto;
-
-public record CreateChallengeCommand(String title,String description) {}
