--- backend/challenge-service/src/main/java/com/itachallenges/challengeservice/application/dto/CreateChallengeResult.java
@@ -1,5 +1,0 @@
-package com.itachallenges.challengeservice.application.dto;
-
-import com.itachallenges.challengeservice.domain.valueobject.ChallengeId;
-
-public record CreateChallengeResult(ChallengeId id) {}
