+++ backend/challenge-service/src/main/java/com/itachallenges/challengeservice/catalog/application/dto/CreateChallengeCommand.java
@@ -1,0 +1,3 @@
+package com.itachallenges.challengeservice.catalog.application.dto;
+
+public record CreateChallengeCommand(String title,String description) {}
