+++ backend/challenge-service/src/main/java/com/itachallenges/challengeservice/catalog/application/dto/CreateChallengeResult.java
@@ -1,0 +1,5 @@
+package com.itachallenges.challengeservice.catalog.application.dto;
+
+import com.itachallenges.challengeservice.catalog.domain.valueobject.ChallengeId;
+
+public record CreateChallengeResult(ChallengeId id) {}
