+++ backend/challenge-service/src/main/java/com/itachallenges/challengeservice/catalog/domain/exception/InvalidChallengeDescriptionException.java
@@ -1,0 +1,8 @@
+package com.itachallenges.challengeservice.catalog.domain.exception;
+
+public class InvalidChallengeDescriptionException extends RuntimeException {
+
+    public InvalidChallengeDescriptionException(String message) {
+        super(message);
+    }
+}
