+++ backend/challenge-service/src/main/java/com/itachallenges/challengeservice/catalog/domain/exception/InvalidChallengeTitleException.java
@@ -1,0 +1,8 @@
+package com.itachallenges.challengeservice.catalog.domain.exception;
+
+public class InvalidChallengeTitleException extends RuntimeException {
+
+    public InvalidChallengeTitleException(String message) {
+        super(message);
+    }
+}
