+++ backend/challenge-service/src/main/java/com/itachallenges/challengeservice/catalog/domain/port/in/CreateChallengeUseCase.java
@@ -1,0 +1,8 @@
+package com.itachallenges.challengeservice.catalog.domain.port.in;
+
+import com.itachallenges.challengeservice.catalog.application.dto.CreateChallengeCommand;
+import com.itachallenges.challengeservice.catalog.application.dto.CreateChallengeResult;
+
+public interface CreateChallengeUseCase {
+    CreateChallengeResult create(CreateChallengeCommand command);
+}
