+++ backend/challenge-service/src/main/java/com/itachallenges/challengeservice/catalog/infrastructure/adapter/in/web/dto/CreateChallengeRequest.java
@@ -1,0 +1,13 @@
+package com.itachallenges.challengeservice.catalog.infrastructure.adapter.in.web.dto;
+
+import com.itachallenges.challengeservice.catalog.application.dto.CreateChallengeCommand;
+import jakarta.validation.constraints.NotBlank;
+
+public record CreateChallengeRequest(
+        @NotBlank String title,
+        @NotBlank String description
+) {
+    public CreateChallengeCommand toCommand() {
+        return new CreateChallengeCommand(title, description);
+    }
+}
