+++ backend/challenge-service/src/main/java/com/itachallenges/challengeservice/catalog/infrastructure/adapter/in/web/dto/CreateChallengeResponse.java
@@ -1,0 +1,5 @@
+package com.itachallenges.challengeservice.catalog.infrastructure.adapter.in.web.dto;
+
+public record CreateChallengeResponse(
+        String id
+) {}
