--- backend/challenge-service/src/main/java/com/itachallenges/challengeservice/domain/exception/InvalidChallengeDescriptionException.java
@@ -1,8 +1,0 @@
-package com.itachallenges.challengeservice.domain.exception;
-
-public class InvalidChallengeDescriptionException extends RuntimeException {
-
-    public InvalidChallengeDescriptionException(String message) {
-        super(message);
-    }
-}
