--- backend/challenge-service/src/main/java/com/itachallenges/challengeservice/domain/exception/InvalidChallengeTitleException.java
@@ -1,8 +1,0 @@
-package com.itachallenges.challengeservice.domain.exception;
-
-public class InvalidChallengeTitleException extends RuntimeException {
-
-    public InvalidChallengeTitleException(String message) {
-        super(message);
-    }
-}
