--- backend/challenge-service/src/main/java/com/itachallenges/challengeservice/domain/port/in/CreateChallengeUseCase.java
@@ -1,8 +1,0 @@
-package com.itachallenges.challengeservice.domain.port.in;
-
-import com.itachallenges.challengeservice.application.dto.CreateChallengeCommand;
-import com.itachallenges.challengeservice.application.dto.CreateChallengeResult;
-
-public interface CreateChallengeUseCase {
-    CreateChallengeResult create(CreateChallengeCommand command);
-}
