--- backend/challenge-service/src/main/java/com/itachallenges/challengeservice/domain/valueobject/ChallengeId.java
@@ -1,25 +1,0 @@
-package com.itachallenges.challengeservice.domain.valueobject;
-
-import java.util.UUID;
-
-public record ChallengeId(UUID value) {
-
-    public ChallengeId {
-        if (value == null) {
-            throw new IllegalArgumentException("ChallengeId cannot be null");
-        }
-    }
-
-    public static ChallengeId random() {
-        return new ChallengeId(UUID.randomUUID());
-    }
-
-    public static ChallengeId from(String raw) {
-        return new ChallengeId(UUID.fromString(raw));
-    }
-
-    @Override
-    public String toString() {
-        return value.toString();
-    }
-}
