--- backend/challenge-service/src/main/java/com/itachallenges/challengeservice/infrastructure/adapter/in/web/dto/CreateChallengeResponse.java
@@ -1,5 +1,0 @@
-package com.itachallenges.challengeservice.infrastructure.adapter.in.web.dto;
-
-public record CreateChallengeResponse(
-        String id
-) {}
